RAG/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
chromadb>=0.4.0
pypdf>=3.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
